Moved BaseStructure into its own module and World into a package
