+++ amulet/api/world/__init__.py
@@ -1,0 +1,1 @@
+from .world import World
